Migrate unit tests to BDD style
